+++ commands/branch_name.go
@@ -5,7 +5,8 @@ import (
 	"fmt"
 	"io"
 	"log/slog"
-	sd "stackeddiff"
+
+	"github.com/joshallenit/stacked-diff/templates"
 )
 
 func createBranchNameCommand(stdOut io.Writer) Command {
@@ -26,7 +27,7 @@ func createBranchNameCommand(stdOut io.Writer) Command {
 				commandError(flagSet, "too many arguments", command.Usage)
 			}
 			indicatorType := checkIndicatorFlag(command, indicatorTypeString)
-			branchName := sd.GetBranchInfo(flagSet.Arg(0), indicatorType).BranchName
+			branchName := templates.GetBranchInfo(flagSet.Arg(0), indicatorType).BranchName
 			fmt.Fprint(stdOut, branchName)
 		}}
 }
