+++ commands/checkout.go
@@ -2,9 +2,10 @@ package commands
 
 import (
 	"flag"
-	sd "stackeddiff"
 
 	ex "github.com/joshallenit/stacked-diff/execute"
+	"github.com/joshallenit/stacked-diff/templates"
+	"github.com/joshallenit/stacked-diff/util"
 )
 
 func createCheckoutCommand() Command {
@@ -15,11 +16,11 @@ func createCheckoutCommand() Command {
 		Summary: "Checks out branch associated with commit indicator",
 		Description: "Checks out the branch associated with commit indicator.\n" +
 			"\n" +
-			"For when you want to merge only the branch with with origin/" + sd.GetMainBranchForHelp() + ",\n" +
-			"rather than your entire local " + sd.GetMainBranchForHelp() + " branch, verify why \n" +
+			"For when you want to merge only the branch with with origin/" + util.GetMainBranchForHelp() + ",\n" +
+			"rather than your entire local " + util.GetMainBranchForHelp() + " branch, verify why \n" +
 			"CI is failing on that particular branch, or for any other reason.\n" +
 			"\n" +
-			"After modifying the branch you can use \"sd replace-commit\" to sync local " + sd.GetMainBranchForHelp() + ".",
+			"After modifying the branch you can use \"sd replace-commit\" to sync local " + util.GetMainBranchForHelp() + ".",
 		Usage: "sd " + flagSet.Name() + " [flags] <commitIndicator>",
 		OnSelected: func(command Command) {
 			if flagSet.NArg() == 0 {
@@ -29,7 +30,7 @@ func createCheckoutCommand() Command {
 				commandError(flagSet, "too many arguments", command.Usage)
 			}
 			indicatorType := checkIndicatorFlag(command, indicatorTypeString)
-			branchName := sd.GetBranchInfo(flagSet.Arg(0), indicatorType).BranchName
+			branchName := templates.GetBranchInfo(flagSet.Arg(0), indicatorType).Branch
 			ex.ExecuteOrDie(ex.ExecuteOptions{Output: ex.NewStandardOutput()}, "git", "checkout", branchName)
 		}}
 }
