+++ commands/command_replace_conflicts.go
@@ -4,6 +4,8 @@ import (
 	"flag"
 	"io"
 	sd "stackeddiff"
+
+	"github.com/joshallenit/stacked-diff/util"
 )
 
 func createReplaceConflictsCommand(stdOut io.Writer) Command {
@@ -14,7 +16,7 @@ func createReplaceConflictsCommand(stdOut io.Writer) Command {
 		Summary: "For failed rebase: replace changes with its associated branch",
 		Description: "During a rebase that failed because of merge conflicts, replace the\n" +
 			"current uncommitted changes (merge conflicts), with the contents\n" +
-			"(diff between origin/" + sd.GetMainBranchForHelp() + " and HEAD) of its associated branch.",
+			"(diff between origin/" + util.GetMainBranchForHelp() + " and HEAD) of its associated branch.",
 		Usage: "sd " + flagSet.Name(),
 		OnSelected: func(command Command) {
 			if flagSet.NArg() > 0 {
