+++ commands/command_update.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/joshallenit/stacked-diff/templates"
+	"github.com/joshallenit/stacked-diff/util"
 )
 
 func createUpdateCommand() Command {
@@ -14,8 +15,8 @@ func createUpdateCommand() Command {
 	reviewers, silent, minChecks := addReviewersFlags(flagSet, "")
 	return Command{
 		FlagSet: flagSet,
-		Summary: "Add commits from " + sd.GetMainBranchForHelp() + " to an existing PR",
-		Description: "Add commits from local " + sd.GetMainBranchForHelp() + " branch to an existing PR.\n" +
+		Summary: "Add commits from " + util.GetMainBranchForHelp() + " to an existing PR",
+		Description: "Add commits from local " + util.GetMainBranchForHelp() + " branch to an existing PR.\n" +
 			"\n" +
 			"Can also add reviewers once PR checks have passed, see \"--reviewers\" flag.",
 		Usage: "sd " + flagSet.Name() + " [flags] <PR commitIndicator> [fixup commitIndicator (defaults to head commit) [fixup commitIndicator...]]",
@@ -28,7 +29,7 @@ func createUpdateCommand() Command {
 			if len(flagSet.Args()) > 1 {
 				otherCommits = flagSet.Args()[1:]
 			}
-			destCommit := sd.GetBranchInfo(flagSet.Arg(0), indicatorType)
+			destCommit := templates.GetBranchInfo(flagSet.Arg(0), indicatorType)
 			sd.UpdatePr(destCommit, otherCommits, indicatorType)
 			if *reviewers != "" {
 				sd.AddReviewersToPr([]string{destCommit.CommitHash}, templates.IndicatorTypeCommit, true, *silent, *minChecks, *reviewers, 30*time.Second)
