--- commands/doc.go
@@ -1,7 +1,0 @@
-/*
-Stacked Diff Workflow.
-
-This is the core functionality of the sd command line interface that is
-implemented in the sd package.
-*/
-package commands
