+++ commands/log.go
@@ -1,18 +1,52 @@
 package commands
 
 import (
-	"fmt"
+	"flag"
 	"io"
+	"log/slog"
+	sd "stackeddiff"
+
+	"fmt"
 	"slices"
 	"strings"
 
 	"github.com/fatih/color"
+
+	ex "github.com/joshallenit/stacked-diff/execute"
+	"github.com/joshallenit/stacked-diff/templates"
 	"github.com/joshallenit/stacked-diff/util"
 )
 
+func createLogCommand(stdOut io.Writer) Command {
+	flagSet := flag.NewFlagSet("log", flag.ContinueOnError)
+
+	return Command{
+		FlagSet: flagSet,
+		Summary: "Displays git log of your changes",
+		Description: "Displays summary of the git commits on current branch that are not\n" +
+			"in the remote branch.\n" +
+			"\n" +
+			"Useful to view list indexes, or copy commit hashes, to use for the\n" +
+			"commitIndicator required by other commands.\n" +
+			"\n" +
+			"A " + color.GreenString("✅") + " means that there is a PR associated with the commit (actually it\n" +
+			"means there is a branch, but having a branch means there is a PR when\n" +
+			"using this workflow). If there is more than one commit on the\n" +
+			"associated branch, those commits are also listed (indented under the\n" +
+			"their associated commit summary).",
+		Usage:           "sd " + flagSet.Name(),
+		DefaultLogLevel: slog.LevelError,
+		OnSelected: func(command Command) {
+			if flagSet.NArg() != 0 {
+				commandError(flagSet, "too many arguments", command.Usage)
+			}
+			sd.PrintGitLog(stdOut)
+		}}
+}
+
 // Prints changes in the current branch compared to the main branch to out.
 func PrintGitLog(out io.Writer) {
-	if GetCurrentBranchName() != util.GetMainBranchOrDie() {
+	if util.GetCurrentBranchName() != util.GetMainBranchOrDie() {
 		gitArgs := []string{"--no-pager", "log", "--pretty=oneline", "--abbrev-commit"}
 		if util.RemoteHasBranch(util.GetMainBranchOrDie()) {
 			gitArgs = append(gitArgs, "origin/"+util.GetMainBranchOrDie()+"..HEAD")
@@ -21,7 +55,7 @@ func PrintGitLog(out io.Writer) {
 		ex.ExecuteOrDie(ex.ExecuteOptions{Output: &ex.ExecutionOutput{Stdout: out, Stderr: out}}, "git", gitArgs...)
 		return
 	}
-	logs := getNewCommits("HEAD")
+	logs := templates.GetNewCommits("HEAD")
 	gitBranchArgs := make([]string, 0, len(logs)+2)
 	gitBranchArgs = append(gitBranchArgs, "branch", "-l")
 	for _, log := range logs {
@@ -36,10 +70,10 @@ func PrintGitLog(out io.Writer) {
 		} else {
 			fmt.Fprint(out, numberPrefix+"   ")
 		}
-		fmt.Fprintln(out, ex.Yellow+log.Commit+ex.Reset+" "+log.Subject)
+		fmt.Fprintln(out, color.YellowString(log.Commit)+" "+log.Subject)
 		// find first commit that is not in main branch
 		if slices.Contains(checkedBranches, log.Branch) {
-			branchCommits := getNewCommits(log.Branch)
+			branchCommits := templates.GetNewCommits(log.Branch)
 			if len(branchCommits) > 1 {
 				for _, branchCommit := range branchCommits {
 					padding := strings.Repeat(" ", len(numberPrefix))
