+++ commands/replace_conflicts.go
@@ -9,17 +9,18 @@ import (
 	"strings"
 
 	ex "github.com/joshallenit/stacked-diff/execute"
+	"github.com/joshallenit/stacked-diff/templates"
 	"github.com/joshallenit/stacked-diff/util"
 )
 
 // For failed rebase: replace changes with its associated branch.
 func ReplaceConflicts(stdOut io.Writer, confirmed bool) {
 	commitWithConflicts := getCommitWithConflicts()
-	branchInfo := GetBranchInfo(commitWithConflicts, IndicatorTypeCommit)
+	branchInfo := templates.GetBranchInfo(commitWithConflicts, IndicatorTypeCommit)
 	checkConfirmed(stdOut, confirmed)
 	ex.ExecuteOrDie(ex.ExecuteOptions{}, "git", "reset", "--hard", "HEAD")
-	slog.Info(fmt.Sprint("Replacing changes (merge conflicts) for failed rebase of commit ", commitWithConflicts, ", with changes from associated branch, ", branchInfo.BranchName))
-	diff := ex.ExecuteOrDie(ex.ExecuteOptions{}, "git", "diff", "--binary", "origin/"+util.GetMainBranchOrDie(), branchInfo.BranchName)
+	slog.Info(fmt.Sprint("Replacing changes (merge conflicts) for failed rebase of commit ", commitWithConflicts, ", with changes from associated branch, ", gitLog.Branch))
+	diff := ex.ExecuteOrDie(ex.ExecuteOptions{}, "git", "diff", "--binary", "origin/"+util.GetMainBranchOrDie(), gitLog.Branch)
 	ex.ExecuteOrDie(ex.ExecuteOptions{Stdin: &diff}, "git", "apply")
 	slog.Info("Adding changes and continuing rebase")
 	ex.ExecuteOrDie(ex.ExecuteOptions{}, "git", "add", ".")
