+++ commands/test_parse_arguments.go
@@ -5,14 +5,14 @@ import (
 	"io"
 	"log/slog"
 
-	"stackeddiff/testinginit"
+	"github.com/joshallenit/stacked-diff/testutil"
 )
 
 func testParseArguments(commandLineArgs ...string) string {
 	panicOnExit := func(stdErr io.Writer, errorCode int, logLevelVar *slog.LevelVar, err any) {
 		panic(err)
 	}
-	out := testinginit.NewWriteRecorder()
-	parseArguments(out, out, flag.NewFlagSet("sd", flag.ContinueOnError), commandLineArgs, panicOnExit)
+	out := testutil.NewWriteRecorder()
+	ParseArguments(out, out, flag.NewFlagSet("sd", flag.ContinueOnError), commandLineArgs, panicOnExit)
 	return out.String()
 }
