+++ commands/wait_for_merge.go
@@ -6,12 +6,13 @@ import (
 	"time"
 
 	ex "github.com/joshallenit/stacked-diff/execute"
+	"github.com/joshallenit/stacked-diff/templates"
 	"github.com/joshallenit/stacked-diff/util"
 )
 
 // Waits for a pull request to be merged.
 func WaitForMerge(commitIndicator string, indicatorType IndicatorType, silent bool) {
-	branchName := GetBranchInfo(commitIndicator, indicatorType).BranchName
+	branchName := templates.GetBranchInfo(commitIndicator, indicatorType).BranchName
 	for getMergedAt(branchName) == "" {
 		slog.Info("Not merged yet...")
 		util.Sleep(30 * time.Second)
